[TD] Removes most use of the "long" data type.
Removes long type which varies in size between supported platforms.
Implements minimum TargetClass needed to deal with TARGET being int.

+++ tiberiandawn/aircraft.cpp
@@ -2556,7 +2556,7 @@ COORDINATE AircraftClass::Target_Coord(void) const
  * HISTORY:                                                                                    *
  *   06/19/1995 JLB : Created.                                                                 *
  *=============================================================================================*/
-RadioMessageType AircraftClass::Receive_Message(RadioClass* from, RadioMessageType message, long& param)
+RadioMessageType AircraftClass::Receive_Message(RadioClass* from, RadioMessageType message, int& param)
 {
     Validate();
     switch (message) {
