+++ tiberiandawn/trigger.h
@@ -137,7 +137,7 @@ public:
     */
     bool Spring(EventType event, ObjectClass* object);             // object-based
     bool Spring(EventType event, CELL cell);                       // cell-based
-    bool Spring(EventType event, HousesType house, long data = 0); // house-based
+    bool Spring(EventType event, HousesType house, int data = 0); // house-based
     bool Remove(void);
 
     /*
@@ -264,8 +264,8 @@ public:
     **	generate the trigger. For time-based triggers, this is the number
     **	of minutes that must elapse.
     */
-    long Data;
-    long DataCopy;
+    int Data;
+    int DataCopy;
 
 private:
     /*
